+++ src/dbms/dbms.cpp
@@ -1,4 +1,10 @@
+#include "gflags/gflags.h"
+
 #include "dbms/dbms.hpp"
+
+DEFINE_string(SNAPSHOT_DIRECTORY, "snapshots",
+              "Relative path to directory in which to save snapshots.");
+DEFINE_bool(RECOVER_ON_STARTUP, false, "Recover database on startup.");
 
 std::unique_ptr<GraphDbAccessor> Dbms::active() {
   return std::make_unique<GraphDbAccessor>(