+++ src/query/engine.hpp
@@ -3,7 +3,6 @@
 #include <experimental/filesystem>
 namespace fs = std::experimental::filesystem;
 
-#include "config/config.hpp"
 #include "data_structures/concurrent/concurrent_map.hpp"
 #include "database/graph_db.hpp"
 #include "logging/loggable.hpp"
@@ -14,6 +13,9 @@ namespace fs = std::experimental::filesystem;
 #include "query/plan_interface.hpp"
 #include "query/preprocessor.hpp"
 #include "utils/dynamic_lib.hpp"
+
+DECLARE_bool(INTERPRET);
+DECLARE_string(COMPILE_DIRECTORY);
 
 /**
  * Responsible for query execution.
@@ -64,7 +66,7 @@ class QueryEngine : public Loggable {
    */
   auto Run(const std::string &query, GraphDbAccessor &db_accessor,
            Stream &stream) {
-    if (CONFIG_BOOL(config::INTERPRET)) {
+    if (FLAGS_INTERPRET) {
       query::Interpret(query, db_accessor, stream);
       return true;
     }
@@ -155,13 +157,13 @@ class QueryEngine : public Loggable {
       return query_plan_it->second->instance();
 
     // find hardcoded query plan if exists
-    auto hardcoded_path = fs::path(CONFIG(config::COMPILE_PATH) + "hardcode/" +
+    auto hardcoded_path = fs::path(FLAGS_COMPILE_DIRECTORY + "hardcode/" +
                                    std::to_string(stripped.hash) + ".cpp");
     if (fs::exists(hardcoded_path))
       return LoadCpp(hardcoded_path, stripped.hash);
 
     // generate query plan
-    auto generated_path = fs::path(CONFIG(config::COMPILE_PATH) +
+    auto generated_path = fs::path(FLAGS_COMPILE_DIRECTORY +
                                    std::to_string(stripped.hash) + ".cpp");
 
     query::frontend::opencypher::Parser parser(stripped.query);
@@ -194,7 +196,7 @@ class QueryEngine : public Loggable {
     // and that is a problem because at this point we want brand new
     // dynamic lib. That is the tmp solution. The right solution would be
     // to deal with this problem in DynamicLib
-    auto path_so = CONFIG(config::COMPILE_PATH) + std::to_string(hash) + "_" +
+    auto path_so = FLAGS_COMPILE_DIRECTORY + std::to_string(hash) + "_" +
                    (std::string)Timestamp::now() + ".so";
 
     plan_compiler.Compile(path_cpp, path_so);