+++ tests/unit/mvcc_gc.cpp
@@ -5,7 +5,6 @@
 #include <memory>
 #include <thread>
 
-#include "config/config.hpp"
 #include "data_structures/concurrent/skiplist.hpp"
 #include "logging/logger.hpp"
 #include "logging/streams/stdout.hpp"